+++ apps/shared/toolbox_helpers.h
@@ -2,15 +2,16 @@
 #define SHARED_TOOLBOX_HELPERS_H
 
 #include <escher/i18n.h>
+#include <string.h>
 
 namespace Shared {
 namespace ToolboxHelpers {
 
 void TextToInsertForCommandMessage(I18n::Message message, char* buffer,
-                                   int bufferSize,
+                                   size_t bufferSize,
                                    bool replaceArgsWithEmptyChar = false);
 void TextToInsertForCommandText(const char* command, int commandLength,
-                                char* buffer, int bufferSize,
+                                char* buffer, size_t bufferSize,
                                 bool replaceArgsWithEmptyChar = false);
 /* Removes the arguments from a command:
  *  - Removes text between parentheses or brackets, except commas */
